Added ability to duplicate behavior states, reorder behavior states, and move behavior states to different categories
Refactored some code related to showing window for adding state and category
Adding category to behavior now saves all elements that have had newly-added categories
